Add files via upload
Module 6 Assignment 

Assignment Instructions: Module 6 - The Transportation 
Model 
Purpose  The purpose of this assignment is to formulate and solve a transportation / transhipment problem. 
In addition, this will help you master the following module outcomes: 
• Formulate a transportation problem.  
• Solve a transportation problem. 
Directions Heart Start produces automated external defibrillators (AEDs) in each of two different plants (A 
and B). The unit production costs and monthly production capacity of the two plants are 
indicated in the table below. The AEDs are sold through three wholesalers. The shipping cost 
from each plant to the warehouse of each wholesaler along with the monthly demand from each 
wholesaler are also indicated in the table. How many AEDs should be produced in each plant, 
and how should they be distributed to each of the three wholesaler warehouses so as to minimize 
the combined cost of production and shipping? 
  
  Unit Shipping Cost Unit Monthly 
  Warehouse 1 Warehouse 2 Warehouse 3 Production 
Cost 
Production 
Capacity 
Plant A $22 $14 $30 $600 100 
Plant B $16 $20 $24 $625 120 
Monthly 
Demand 
80 60 70     
Formulate and solve this transportation problem.